requirements.txt
bleach==1.5.0
click==6.7
enum34==1.1.6
Flask==0.12.2
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.2
Markdown==2.6.10
MarkupSafe==1.0
numpy==1.13.1+mkl
olefile==0.44
Pillow==4.3.0
protobuf==3.5.1
PyYAML==3.12
scipy==0.19.1
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.13
